initial project restructure
