+++ build.gradle
@@ -1,21 +1,20 @@
 apply plugin: 'java'
 
-
 repositories {
     mavenCentral()
 }
 
 dependencies {
     compile group: 'com.google.code.gson', name: 'gson', version: '2.8.0'
-    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.9'
-    compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.8.9'
     compile group: 'net.portswigger.burp.extender', name: 'burp-extender-api', version: '1.7.22'
+    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.8.9'
+    compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.8.9'    
 }
 
 sourceSets {
     main {
         java {
-            srcDir 'burp'
+            srcDir 'src'
         }
     }
 }
