+++ src/swurg/Scheme.java
@@ -14,7 +14,7 @@
 #    limitations under the License. 
 */
 
-package burp;
+package swurg;
 
 public class Scheme {
     private int port;
